[arithmetic][PotentialFix] ArgShiftRule handle stacking dim.
In stochastic, the asloss method is used for l2norm. I have a 3 channel image problem and need to pass a ground truth vector
in the asloss() that is (3, xxx). ie has a stacking dimension.

To allow this I deleted the requirment that the shifting constant be a 1D vector, and change the dim calculation to look
at the last dimension in the vector. To me it makes sense that the ground truth data should be flattened and can contain stacking dimensions.