[NUFFT] Partial revert of 8b3d8bb and 7369266 (see notes).
The presence of the upsampling factor in the multiplicative scaling of the target meshgrid was actually correct my bad. The other mistakes are actually inconsistencies between the FINUFFT paper and package. I revert to give precedence to to the actual implementation over the paper as Sepand had originally done (in the end this is what matters). I noted the inconsistencies in the code.