[op][stencil] Lift precision-specific constraint (with warning).
Stencils only work with input dtypes which match the kernel's.
We purposely do NOT want to force a copy for performance reasons.
However, it is common for novice users to not take care of precision issues, thus stumble when using Stencil().

Decision: allow using alternative dtype-inputs, but warn user if sub-optimal casting required.