[fft] Discontinue pyFFTW backend.
Reason: SciPy/CuPy versions work across process-based DASK workloads.
PyFFTW on the other hand needs to create a plan object which cannot be
passed to other processes.

Decision taken to discontinue it from the API.
It is preferable to create an extension module based on ducc0 for those who want added performance.