[util] Add require_viewable().
view_as_real/complex() require last-axis contiguity (for some backends) to work.
It is the responsibility of the user to guarantee this manually.
However, sometimes some operators are used as sub-operators in other operators, in which case we want to ensure the sub-operator works without having to test on a case-by-case basis.
require_viewable() refactors this logic into a re-useable building block for these cases.