[SWE] Add opt-einsum as core dependency.
This is required to perform efficient multi-dimensional contractions and/or broadcasting in fft/ and perhaps beyond.
opt-einsum has no dependencies other than NumPy, and is heavily used in other libraries; as such there is a high change we are pulling it already.
This commit makes the dependency explicit, without affecting portability across operating systems.